refactor: Misc refactoring
